requirements.txt
Django
psycopg2
django-phonenumber-field
sk-video
av
pyparsing
djangorestframework
